requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
coloredlogs==15.0.1
    # via onnxruntime
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
evdev==1.9.2
    # via pynput
filelock==3.18.0
    # via torch
flatbuffers==25.2.10
    # via onnxruntime
fonttools==4.59.0
    # via matplotlib
fsspec==2025.7.0
    # via torch
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via requests
jinja2==3.1.6
    # via torch
kiwisolver==1.4.8
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.3
    # via ultralytics
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
mss==10.0.0
    # via -r requirements.in
networkx==3.4.2
    # via torch
numpy==2.2.6
    # via
    #   contourpy
    #   matplotlib
    #   onnx
    #   onnxruntime
    #   opencv-python
    #   pandas
    #   scipy
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
onnx==1.18.0
    # via
    #   -r requirements.in
    #   onnxsim
onnxruntime==1.22.1
    # via -r requirements.in
onnxsim==0.4.36
    # via -r requirements.in
opencv-python==4.12.0.88
    # via
    #   -r requirements.in
    #   ultralytics
packaging==25.0
    # via
    #   matplotlib
    #   onnxruntime
pandas==2.3.1
    # via ultralytics
pillow==11.3.0
    # via
    #   matplotlib
    #   torchvision
    #   ultralytics
pillow-simd==9.5.0.post2
    # via -r requirements.in
protobuf==6.31.1
    # via
    #   onnx
    #   onnxruntime
psutil==7.0.0
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pygments==2.19.2
    # via rich
pynput==1.8.1
    # via -r requirements.in
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-xlib==0.33
    # via pynput
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via ultralytics
requests==2.32.4
    # via ultralytics
rich==14.1.0
    # via onnxsim
scipy==1.15.3
    # via ultralytics
setuptools==80.9.0
    # via triton
six==1.17.0
    # via
    #   pynput
    #   python-dateutil
    #   python-xlib
sympy==1.14.0
    # via
    #   onnxruntime
    #   torch
torch==2.7.1
    # via
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchvision==0.22.1
    # via ultralytics
tqdm==4.67.1
    # via ultralytics
triton==3.3.1
    # via torch
typing-extensions==4.14.1
    # via
    #   onnx
    #   torch
tzdata==2025.2
    # via pandas
ultralytics==8.3.169
    # via -r requirements.in
ultralytics-thop==2.0.14
    # via ultralytics
urllib3==2.5.0
    # via requests
